split build graph json file read effect tests
